+++ rsa.c
@@ -4,8 +4,8 @@ int main(int ac, char **av)
 {
 	FILE *fp;
 	char line[1024];
-	long long int num;
-	long int f_prime1, f_prime2;
+	ull num;
+	ull f_prime1, f_prime2;
 	int i;
 
 	if (ac != 2)
@@ -25,16 +25,16 @@ int main(int ac, char **av)
 		num = strtoll(line, NULL, 10);
 		f_prime1 = find_factor1(num);
 		f_prime2 = find_factor2(num, f_prime1);
-		printf("%lld=%ld*%ld\n", num, f_prime2, f_prime1);
+		printf("%llu=%llu*%llu\n", num, f_prime2, f_prime1);
 	}
 	fclose(fp);
 	return(0);
 		
 }
 
-int find_factor1(long long int num)
+ull find_factor1(ull num)
 {
-	long int f1;
+	ull f1;
 
 	for(f1 = 2; f1 < num; f1++)
 	{
@@ -43,9 +43,9 @@ int find_factor1(long long int num)
 	}
 	return (0);
 }
-int find_factor2(long long int num, long int f1)
+ull find_factor2(ull num, ull f1)
 {
-	long int f2;
+	ull f2;
 
 	f2 = num / f1;
 
@@ -55,7 +55,7 @@ int find_factor2(long long int num, long int f1)
 	return (0);
 }
 
-int check_prime(long int n)
+int check_prime(ull n)
 {
 	if (n < 2) 
         	return (0);
